feat(runtime): add runtime configuration options for caching and programmatic control
