+++ crates/runtime/tests/cluster/job_store.rs
@@ -119,6 +119,7 @@ async fn test_simple_job_store() -> Result<(), anyhow::Error> {
                     node_mtls_key_file: Some(scheduler_cert.key_path.to_string_lossy().to_string()),
                     ..Default::default()
                 },
+                ..Default::default()
             };
 
             let scheduler_rt = Arc::new(
@@ -176,6 +177,7 @@ async fn test_simple_job_store() -> Result<(), anyhow::Error> {
                     node_mtls_key_file: Some(executor_cert.key_path.to_string_lossy().to_string()),
                     ..Default::default()
                 },
+                ..Default::default()
             };
 
             let executor_rt = Arc::new(
